+++ projects/ngx-editor/src/lib/Locals.ts
@@ -21,6 +21,8 @@ export const defaults: Record<string, string> = {
   align_justify: 'Justify',
   text_color: 'Text Color',
   background_color: 'Background Color',
+  horizontal_rule: 'Horizontal rule',
+  format_clear: 'Clear Formatting',
   insertLink: 'Insert Link',
   removeLink: 'Remove Link',
   insertImage: 'Insert Image',