+++ projects/ngx-editor/src/lib/modules/menu/MenuCommands.ts
@@ -25,6 +25,7 @@ export const ToggleCommands: Record<string, ToggleCommand> = {
 
 export const InsertCommands: Record<string, InsertCommand> = {
   horizontal_rule: Commands.HORIZONTAL_RULE,
+  format_clear: Commands.FORMAT_CLEAR,
 };
 
 export const Link = Commands.LINK;